fix: miden executable errors
